requirements.txt
flask
python-dotenv
opencv-python
Pillow
azure-functions
azure-storage-blob